move layer proving info structs to util file
